(server) Improve main game algorithm
Now a tile strength is a virtual / dynamic attribute that must be
computed. We only store a "damage" counter that will help determine if
a user can take a tile's ownership. See `GameData::computed_tile` and
`GameData::handle_click` for more details on the implementation.

This simplify computing quite a lot compared to previous implementation
and makes it more consistant too. There was some "gaps" and incoherences
with previous implementation.