(server) Enforce core mechanics with new tests
Add new tests to enforce core game mechanics which should now be fully
covered. All grid tile update mechanism is covered.

Also lighten the way to handle the grid and remove potentially unowned
tiles (aka tiles with a user_id = None). This make the overall logic
more simple and clear even if it adds a little bit of logic. Greatly
simplify management and comparison of `user_id` tile attribute and also
reduce the size of the grid by default.

With those new tests & the new way to handle the grid state I hope it
will help future updates of the surounding logic like how to retrieve
the grid tiles and how we store them.

My current plan at this moment in the modifications is to implement a
batch system to fetch the grid state in multiple steps by:
- Removing `GET /data` endpoint.
- Adding a new `GET /users` endpoint. This implies that we change the
  score system. Which is not a big deal since it's not really complete.
- Adding a new `GET /tiles?batch={batchNumber}` endpoint. So we need a
  way to create batches of coordinates that split the grid in `N`batch-
  es of `S` size.

Once that done we should benchmark how this behaves in heavy load via
locust. Ideally we should be able to fetch the full grid state without
crashing the app (which is the case now).

Depending on the locust results we may have to plan a better horizontal
scaling system by changing the way of storing the grid via redis and use
kubernetes. This would allow us to have multiple instance of actix-web
acting on and reading from the same grid shared via redis.

The final goal being to know how much maximum request per second the app
can take and what policy we should take concerning the number of concur-
rent users.