Start to implement batch system + fix front draw
Made huge progress for batch system by succesfully paved the grid in a
more or less balanced pattern. Not perfect should suffice to implement
final batch system in later modifications.

Also debugged the way the grid was drawn. It was inverted, making me
think the batch algorithm was broken when in fact it was working as
expected.