[experimental] optimize fetching neighbors
Does not seem mush but if I'm not wrong the small modifications I made
here should optimize the way we process the various tiles strength
computation by reducing access to redis by handling those fetched tiles
in a temporary external hashmap (external from the part doing the fetch
and computation).

This can produce small inconsistency if we look closely the data on an
millisec level. But do be honnest this probably was true with previous
implem.

However it seems, for an unexplainable reason that this change broke
locust tests. Maybe not this modification particularly, needs debug.